Reorganized and renamed files to prevent import errors when used as submodule.
Having an additional folder caused import errors when used as a submodule. This
was the best fix. Also, removed the old pygatt code because it was causing tox
errors. Will add in when time to refactor it.
