+++ pygatt/bled112_error.py
